New monster: Ironwhip Beastmaster
Instead of having beasts spawn in the vaults on their own, let's add
on a Vaults thematic strength: bigger fights with synergizing bands. The
Ironwhip Beastmasters are the ones who have taken the late, rare Lair
monsters to the Vaults.

Beastmasters come with a "Goad Beasts" natural ability, which gives all
beasts in LOS a bonus to energy equal to one EUT_MOVE. This is less
pronounced than haste other or a torpor snail aura for basic combat;
however they come with bands of trampling dire elephants or fire
breathing lindwurms.

Here the goad helps the monsters in the band manouver around one another
to better get to the player. For trampling monsters this does have a
different mechanical impact than a torpor snail aura: after a successful
trample that moves the player the remaining goaded elephants have a
better chance at getting a hit.

Beastmasters also spawn with nets, because what's a good beast master
without a net.