New spell: Pandemonium Form (L9 Transmutations)
If you manage to get it in 3 runes, it'll carry you pretty nicely for a win. The form has large size, 35% GDR, 10 base AC with up to 6 spellpower-based extra AC, large size, 13 base UC damage and chaos-branded UC, +30% HP, rPois, rElec, flight, and randomly determined resistances to cold and fire (these can be vulnerabilities too, and the form melds all worn items, so you can't rely on your jewelry). The form also gives you demonic holiness and a set of random abilities - it can pick from some of the better draconian breaths, blinking, and/or damnation identical to the demonspawn mutation. form is also uncancellable, so you can't spam-cast it for good resistances and abilities. Chaotic spell, so good gods hate this. This replaces Haunt in the Grand Grimoire, and that spell has been moved to the Necronomicon.